Auto-add a review date to new capture to-dos
... for the next day. This seems reasonable under the assumption that
one would add a capture to today's list instead of the capture, if it
was expected to be finished earlier. It also means that one can add a
new item to the capture list and restart Habituary (on the same day),
without having to review the to-do one just added.