Update a to-do's review date in `_drop_data`
If a to-do from the capture list is dragged to any other list, its
review date (and those of its sub items, if there are any) will be
removed. If a to-do from any day is moved to the capture list, its
review date is updated to equal the date of tomorrow.

Also, reduces code redundancy by calling the sub item lists `_drop_data`
function in the else-case of the `_drop_data` function in todo_item.gd.

Further, simplifies yesterdays commit (c22a04d), by making use of the
to-do item's `review_date` variable instead of setting it's `text`.