Fix several smaller issues with the review system
The tooltips when hovering over a DayButton of the calendar widget can
now be customized via an export variable of the widget. "Click to jump
to this date" still remains the default, but both the calendar widget
for postponing and for scheduling to-dos use a custom variant now.

Scheduling a reviewed to-do for the same day is now possible again. It
was blocked by an if-condition, that required a day difference of more
than zero – however, that only makes sense for postponing!

The `_schedule_current_item` function now uses `Cache.move_item` again,
effectively undoing 2c9cd42. While that approach had its benefits, too,
it didn't feel right to me that the original remained in the list. As an
aside, it's easier to explain what "scheduling" does that way!

The `save_to_disk` function in `Cache` will now delete the entire key if
the key's `content` variable is empty. Without that, scheduling the last
to-do in the capture list looked right, but the changed capture list was
not stored, resulting in another review after restarting the app.

Adding the review date to newly added to-dos in the capture list is now
part of the setter function for the to-do's text, to ensure it's always
done. Previously it could be skipped when exiting the app while editing
a to-do from the capture list before it got submitted.