Rewrite calendar widget's `update_day` method
This fixes a regression which resulted in an error when changing the
variable `Settings.current_day` while a calendar widget was visible
where `include_past` was set to false. Since both calendar widgets in
the review view remain technically visible (only one of their parents is
hidden, cf. `is_visible_in_tree`) that was a real issue in practice! The
new implementation is much cleaner and fully robust to this issue.