Fix some more theming issues
Unifies the appearance of the editing options:
- Fixes `EditingOption` and `ToDoItem_BookmarkButton` not having a
  distinct color on mouse hover events while `dark_mode` is `true`.
- Adds a similar hover effect for the `TextColor` panel.

Further, adds a new `CollapsibleCategory_Heading` theme type that
differes from the regular button style in two mayor ways:
- The `bg_color` remains `primary_3` (blue), independent of the state.
- The font colors are chosen in such a way that `NEUTRAL_2` (black) is
  used while the category is folded, `NEUTRAL_7` (white) is used while
  it's unfolded and `secondary_3` (orange) on all hover events.

Also, *all* categories in the settings are now folded by default. This
makes more sense if the user is lookign for a specific category, since
they don't have to first close the "General" category on small screens.