Fix tooltip appearance
Previously there was no horizontal space between a tooltip's text and
its shortcut (if it had one), as the default BoxContainer `separation`
in the theme is now set to 0 and no theme type variation was provided.

Also, the shortcut text will no longer be smaller than the tooltip's
text, which looked especially weird in the dense configuration.