Fix wrong vertical scroll value after review
When restoring a to-do list with enough items to be scrollable and
scrolled to the last item, the last item (a) wasn't visible and (b)
couldn't be scrolled back into view, until either the DayPanel got
reloaded or another to-do was added to the list.

I'm not sure what causes this, but it only happens after the review
view - if capture reviews are disabled, it doesn't occur! Through
experimentation, I found out that waiting one process frame before
switching the main panel to the capture review prevents this, too.
Curiously, using `set_deferred` to change the main panel state does
*not* achieve the same result. If I had to guess, I would say that
something in the scrollable todo list isn't initialized correctly
when changing it's (indirect) visibility too early.

Anyways, this simple change seems to fix the bug!