Fix bugs after scheduling more than one item
Since scheduling a reviewed item will move it *out* of the capture list,
and the line ids used to specify which item should be moved are gathered
in advance at the start of the reviewing process, scheduling multiple
items in one session let to either (a) moving the wrong item or (b)
trying to move a non-existent item with a too high line id.

This commit adresses this by introducing a `scheduled_items` counter and
subtracting it's value from the line id of the item to be moved. As long
as there's no way to review the items out of order, or edit the capture
list otherwise, this should fix the problem.