Add a "Copy to Today" button to past to-dos
The button won't be available for to-dos due today or in the future! It
is meant as an easy and quick way to copy over to-dos that are supposed
to be repeated today, either because they couldn't get accomplished last
time, or simply because they are part of a regular routine.

It sits at the left side of the drag handle, and once pressed, will copy
the to-do over to today. It does *not* matter if today is visible in the
list view! Naturally, to-dos with no text cannot be copied over.

This commit also adds a new EventBus signal `instant_save_requested`. It
prompts all visible to-do lists to immediately save all pending changes.
This is required to prevent race conditions with the debouncer, e.g.:
- Copying over a newly created, currently edited to-do (which doesn't
  yet exists in the cache)
- Deleting to-dos from today, then copying another to-do over (before
  the debouncing timer for saving the deletions to cache runs out)