Add a "Saved Searches" side panel
This is a work in progress, eventually meant to replace the bookmarks
panel. For the moment, it has no shortcut assigned, and only works when
one creates a `saved_search.txt` file outside of Habituary. There is no
option to add or remove a saved search query in the app (yet)!

The saved searches are displayed similarly to the bookmarks, but their
`date` variable equals the date of the latest to-do matching that query,
and the button next to it triggers a global search for the query.
Like the bookmarks the list is sorted by item date, with earlier dates
being higher up. Queries with no matching to-dos will be displayed at
the top (as - without matches - they're are likely to get deleted).

Also, adds a MIN_INT and MAX_INT constant to `Utils` for convenience.