Add an option to add or remove a saved search
When doing a global search, there is a new button now, that either reads
"Save This Search" or "Unsave This Search" (when the respective query
already has been added to the list of saved searches). This means it's
now possible to populate and clear the list from inside the app!

Also, modifies the text of the "NoneSet" label, to indicate to the user
how they can add a query (since the panel initially will be empty).

To make all of this work, the `saved_searches.txt` file is now included
in the list of cached files, and both `update_content` and `delete_key`
now accept a second, optional parameter `emit` that defaults to `true`,
and will cause the `content_updated` signal to be emitted. (Note that
this behaviour is *not* wanted when the debounce timer of an mofidied
to-do runs out, as that would reload the respective to-do list!)