Update saved search results on to-do text changes
That is, when a to-dos text is either changed to or from a text matching
the query of one or multiple saved search queries, or if such a to-do is
either deleted or moved to another list.

This commit introduces a new `saved_search_update_requested` signal, and
an `_update_saved_search_results` method in `todo_item.gd` that emits it
when necessary. The `saved_item.gd` script then reacts to this signal by
calling `_find_last_mention` again each time the signal is emitted.

Also, fixes an oversight when iterating over the all the cached keys in
descending order: Just calling `cahe_keys.reverse()` wasn't enough when
dealing with situations where a new key was added during runtime (e.g.
by creating a to-do on a previously empty date). So now the list of keys
gets properly *sorted* n descending order each time.

Furthermore, makes sure to call `_update_copy_to_today_visibility` when
moving a to-do to another (potentially future or today) date now!