Show a toast notification when copying a to-do
This commit adds a new `Overlay` autoload with a `spawn_toast` function,
that expects a string as its argument and will spawn a notification with
said string as its text at the bottom center of the window. The overlay
*does* take into account whether the side panel is open or not!

By default, a toast will slowly fade out and close after 2 seconds. This
can be sped up by clicking on it, which will fade out in 0.3 seconds.

If multiple toasts spawn in quick succession, they'll vertically stack.