Add optional ALARM tag for saved searches
If set, the tag provides either a negative or positive integer value as
a `warning_threshold` for a saved search. When that item's day counter
is updated, and its `day_diff` is smaller or equal to said threshold,
the font color of its `DayCounter` will turn red. This way it's possible
to set up a visual warning if a future to-do comes near, or a past to-do
hasn't been repeated in a while (useful for habit tracking).

There currently is *no* way to set an alarm in the app: One has to edit
the `saved_searches.txt` file manually for now!