Add UI for adding or removing a saved search alarm
This should be fully functional for now, but isn't yet saved to disk, so
it won't persist across restarts. There's a new bell shaped button next
to the DayCounter of every saved search, that can be clicked to unfold a
third line for said item allowing to set both the number of days as well
as whether this alarm should trigger before or after the event. Clicking
the bell shapes button again will remove the alarm.