Make alarm changes persist across restarts
As mentioned in the previous commit, modifying the alarm settings of a
saved search wasn't yet saved to disk. This commit changes that! Analog
to the to-do lists, saves are debounced for 10 seconds or enforced when
either quitting Habituary or focusing another app.

Also, I realized that  `NOTIFICATION_WM_WINDOW_FOCUS_OUT` triggers when
e.g. just opening a drop down menu as well (regardless of whether or not
the "Embed Subwindows" project setting is active or not), so this commit
replaces mentions of that with `NOTIFICATION_APPLICATION_FOCUS_OUT`.

Furthermore fixes a rather strange bug when calling `_update_third_row`
after resetting the `warning_threshold` to `Utils.MIN_INT`: It sets the
value of the SpinBox to the absolute value of the threshold, which in
this case happens to be `MIN_INT` again. This gets further obfuscated by
the spin box having a `min_value` of 1.0, so that's what the connected
`on_value_changed` callback receives. It then proceeded to reset the
`warning_threshold` to that value and calling `_update_third_row` again,
leading to the wrong (red) font color when removing an alarm.
The fix is trivial, though: Simply don't update the SpinBox and Option-
Button when the `warning_threshold` equals `Utils.MIN_INT`.