Request list save when moving a to-do via keyboard
Previously, calling `move_todo_up` and `move_todo_down` both did *not*
start the list's debounce timer, which let to setbacks when moving one
of today's to-dos and then copying another to-do to today.

The debounce timer will not be started when moving newly created to-dos
(i.e. those where the `text` variable is still empty), since those are
automatically saved when submitting the to-do later on anyway.