Include sub items when copying an item to today
This is a bandaid fix, until I find the time and motivation to rework
the `Cache` autoload to make it easier to get all sub items of a to-do
right there, without the need to loop over successor lines each time.
With this it's now possible to copy a to-do AND all of its sub items
(including their sub item's sub items) to today, keeping their relative
indentation intact, while removing all indentation of the base to-do.