Release focus on to-do text submission
As noted in the previous commit message (02876a2), I thought the call to
`%Edit.release_focus` was no longer needed. However, it *is* needed when
editing an already existing to-do, as just hitting ENTER won't hide the
EditingOptions otherwise. This commit fixes that, while making sure the
function is only called while `key_input` is true to prevent the errors
when outside the tree that I tried to fix in the previous commit.