Make relative saved search dates optional
... by adding a new "Use Relative Dates" setting, and using absolute
dates instead if the user sets it to `false`. This is a purely visual
change, the highlighting logic for set alarms doesn't change!

Also, adds a "Latest Match:" label in front of the `DayCounter` to
clarify what the date refers to, and changes the DayCounter's text
overrun behavior to "Ellipsis", just in case.

This setting *does* persist across restarts.