Fix showing tooltip after the host got hidden
This was the case for the CalendarWidget, for example, where clicking on
a DayButton would hide the widget, but also trigger `show_tooltip` anew.
Once the `_hover_timer` ran out, the tooltip would spawn again, and only
disappear once the user clicked anywhere or typed something, the regular
way of hiding it on the host's `mouse_exited` signal did *not* work!