Don't require weekday for creating a Date anymore
The `weekday` can be (and in fact: already was) derived from the other
properties `year`, `month` and `day`, so explicitly specifying it when
creating a new `Date` object is no longer required now.

To keep the class compatible with the dictionaries returned by Godot's
`get_date_dict_from_system` method, the `_init` method only checks if
the dictionary contains a "year", "month" and "day" key and will ignore
all other keys. The `weekday` variable is automatically updated when the
`day` setter runs, which happens whenever the `Date` is updated.