Fix Escape key not working in Capture panel
The bug resulted from the to-do item's `delete` method throwing an error
because it was trying to use `date.as_string()`, to get the cache key of
the item in order to update the saved search results. Since items in the
capture list don't have a `date` this didn't work. Instead the code now
gets the cache key by directly querying the superior to-do list.