Stop tooltips from showing while dragging items
... unless explicitly enabled via the tooltip's `show_during_dragging`
export variable. With this becoming the default behavior, a passge in
`day_panel.gd` achieving the same result is now superfluous.

Also, fixes the to-do item's `UnfoldTimer` starting the debounce timer
when hovering a dragged item over an *already* unfolded item, by having
the `is_folded` setter return early if the value is already set.