Fix to-do in edit mode disappearing on list save
This effectively turns the second change in a6bab83 into a general part
of the to-do's `as_string` method. Without this, it was possible to e.g.
move the last to-do in a list to another date (or delete it), then start
editing a new to-do in that now empty list, and not submitting it's text
until the debounce timer ran out, triggering `save_to_disk` and deleting
the edited to-do in the process (as it got ignored by `as_string`).

Also, stopping the DebounceTimer is now part of `save_to_disk`. For an
instant save request this might be required to prevent saving twice (if
the timer was already running when the request arrived), for all other
cases stopping the timer should make no difference (it's stopped).