Resort saved searches if match date changes
Previously the order would only update when `load_from_disk` was called,
but remain as it was if the date of the last matching item changed (e.g.
because an item with a matching text got added or deleted).

When `load_from_disk` is called, old items are now immediately free'd
(instead of using `queue_free`) to reduce the number of items to sort.
Furthermore, `_resort_list()` will be skipped now if the number of items
is less than two (since there's nothing to sort then).

Also, fixes a bug introduced by the previous commit 7c11eea: Turns out
that toggling the EXTEND flag of the "Add Alarm" button was *not* the
right solution, as this will steal space from the DayCounter.
The DayCounter text is now set to the empty string instead of hiding it
when there's no date set, so its own EXTEND flag remains active.