Fix orphaned nodes remaining on exit
Adds a few lines of code to `Utils.gd` that will print a warning in red
if there are orphan nodes remaining when quitting the app.

Turns out, adding a scene instance to a variable in the global scope of
a script isn't particularly clever if you want to avoid orphans... The
reason for this change stated in 2417ff4 no longer applies (likely since
introducing the Cache fixed it), so I've reverted to show & hide.

Also, fixes a bunch of issues with to-do's in the capture panel, because
the code was still expecting all to-dos to have a date  – but to-dos in
the capture panel don't!. Instead the list's `cache_key` is used.