Auto-add sub item for edited items ending in colon
For freshly added items this was already the case. However, when editing
an already existing item and adding a colon in the end, no sub item was
added below it unless manually triggered (by pressing SHIFT+ENTER). This
commit changes this and will automatically add a sub item if...
- ... the item does not have any sub items yet, and...
- ... the edited text ends with a colon while the previous text doesn't.

Also, fixes a bug in the EditingOptions that would only occur when the
mouse cursor was hovering over the `TextColor` button when the focus of
the respective to-do was released (e.g. by submitting or deleting it),
by returning early if the item's editing options are already queued for
deletion when the `_on_dark_mode_changed` callback is run.