Add `Data` singleton (WIP)
This is meant to eventually replace the `Cache` singleton, and aims to
address several shortcomings of its current implementation. The goal is
for *all* data manipulation to happen inside this node, while all other
scenes will merely render a "view" of that centralized data.

For now, both Singletons will co-exist, until the new `Data` singleton
step by step gets ready to replace it's predecessor.

This first draft represents each to-do list in `Settings.store_path` as
a RefCounted script, which represents each of its items as a RefCounted
script, too. Empty files and/or lines will be ignored.
Note that indented to-do's aren't implemented yet!