Reorganize data types
Both by adding a new one, and assigning them class names.

What was previously called `todo_list.gd` is now called `file.gd`, and
can be accessed as the `FileData` class. If there are any valid to-dos
in the file, the (new) `ToDoListData` class gets created to hold them.
Therein, each to-do is represented by a `ToDoData` class.

This is meant as a preparation step for properly integrating indented
to-dos into the `Data` singleton as well. By separating `ToDoListData`
and `FileData`, it should be possible to assign to-dos with sub items
their own `ToDoListData` class to hold those sub items.