Add support for indented to-dos
This is mostly analogous to the implemenetation in `todo_item.gd`. It
required a slight change to how ToDoData objects are initialized to make
sure `load_from_string` (previously called `_parse`) isn't called before
the `parent_list` variable has been set: we first create the object, add
it to its to-do list (which automatically assigns the `parent_list`) and
only *then* load the to-do's properties from the raw string.

There's a slight overhead for indenting a to-do multiple levels deep (as
it will be indented one level at a time and removed and added to a to-do
list array each time) that might be avoidable with a different approach,
but since most to-dos won't be indented at all, I think that's okay.