Refine `_valid_filename_reg_ex`
The previous version did also accept filenames that contained arbitrary
strings before and/or after the reg ex match. It also didn't check if
the month was valid (i.e. not "00" or anything bigger than "12"). Same
with the day, except here one also has to account for leap years and the
various month lengths (i.e. 28, 29, 30 or 31 days).