Automatically save to disk if a file gets changed
Consider this a proof of concept: It still has some issues and isn't yet
used anywhere productively in the app!

Changing any property of a to-do (that is: `state`, `text`, `is_bold`,
`is_italic`, `is_bookmarked`, `is_folded` or `text_color_id`) emits a
`changed` signal, which gets forwarded by its parent to-do list. Also,
if a to-do is added to or removed from a to-do list, the list will emit
its own `changed` signal as well. (Which leads to the first issue: Some
changes might result in multiple emissions of the `changed` signal!)

Once the signal reaches the `FileData` it will kick off a debounce timer
of 10 seconds, which will call a new `_save_to_disk` method on timeout.
Since the `Timer` node only works when added to the tree, and RefCounted
objects are deliberately *not* part of the tree, setting up the debounce
timer turned out to be more difficult than I had anticipated.
To solve this, I ended up letting `FileData` extend `Timer` and added it
as a child of `Data`, effectively replacing the old `files` dictionary.
While that works, it now dawned at me that it might have been better to
simply let the `Data` singleton set up a debounce timer in its stead and
keep the `FileData` as `RefCounted` objects. (Considering it's unlikely
that all loaded files will simultaneously run a debounce timer).

Also, `_notification` now uses `NOTIFICATION_APPLICATION_FOCUS_IN|OUT`
instead of `NOTIFICATION_WM_WINDOW_FOCUS_IN|OUT` (cf. 65e2447).