Let `Data` create the the debounce timers
As described in the last commit (c01ba25), this makes more sense than
the previous approach, as it's unlikely that all the loaded files will
require a debounce timer at the same time.

To prevent `FileData` from triggering a debounce timer when `reload` is
called, the callback gets disconnected during `_load_from_disk`.

Adds a `name` property to the `FileData` for easier access.