Make search results with long to-do texts wrap
Theoretically, this could have been as simple as activating any autowrap
mode on the RichTextLabel and setting its horizontal alginment to expand.
However, since the ExtraPadding's horizontal alignment is already set to
expand (and required for it to function properly!), the solution is more
complex and involves calculating a `strech_ratio` for both nodes.

The ExtraPadding gets assigned a `custom_minimum_size` of 15 pixels to
guarantee there always will be at least three visible dots.

Also, changes the vertical alignment of the State, Text, ExtraPadding,
and JumpTo nodes to "Shrink Begin", and adjusts the size of the JumpTo
button to match the size of the State icon.