Fix orphaned `seach_result_group` on app exit
I was under the impression the instantiated `search_result_group` would
be automatically free'd at the end of the function scope as long as one
didn't add it to the tree. Turns out, that's not the case!

Also, moves the code in `Utils` that checks for orphan nodes to its own
method, so it can easily be called during runtime as well, and includes
a call to the built-in `print_orphan_nodes` method to it, which gives a
better indication about *which* nodes end up orphaned. (Only available
in debug builds. In released projects it won't do anything.)