Update hover indicators on `item_rect_changed`
When e.g. moving a to-do via a keyboard shortcut or resizing the window,
it can happen that the mouse cursor no longer is ontop of the CheckBox
and/or DragHandle. However, the visuals would still indicate that until
the user moves the mouse. This commit enforces an inmediate update!