Avoid memory leak due to cyclic dependency
Previously, each to-do contained a reference to it's parent list, which
contained a reference back to the to-do. Thus, both the to-dos and the
lists didn't get properly free'd when exiting Habituary.

I *could* have solved this by turning `parent_list` into a `WeakRef`,
but it felt rather verbose (as you always have to include `get_ref()`
then and typed WeakRefs aren't yet supported, see:
https://github.com/godotengine/godot-proposals/issues/9174

So instead, I got rid of the back reference altogether! To-dos now emit
a signal, if they want to request deletion or an indent, which then is
connected to the appropriate callbacks by the to-do list.
For the `indentation_level` that approach doesn't work, as the info is
needed in the to-do's script itself. Thus, the respective variable has
to be duplicated from the one in the parent for each to-do.