Restructure Help panel using FoldableContainers
The sections mostly remained the same. The "WHILE EDITING A TO-DO" one
was moved to the beginning of the list, as for most users it's likely
the most relevant one. `toggle_fullscreen` and `toggle_dark_mode` were
moved to their own section ("CHANGE LOOK"). Also, the wording of some
descriptions has been shortened where the context allowed for it.