Fix `stretch_ratio` calculation for search results
The old approach (introduced in 2b5d9fd) used the `custom_minimum_size`
of both the "State" indicator and the "JumpTo" button. However, during
the switch to DPITextures (see 7a0b04d) those values got nulled again.
So instead we now have to use the texture width for correct results!

Also, rearranges the order of subtrahends in the `max_width` calculation
to match their position in the scene tree (from top to bottom).