Fully migrate the overlay into an autoload
Previously, the `main_window` still contained an "Overlay" scene that
handled the calendar widget opening and closing. Instead, this is now
handled by the "Overlay" singleton - which already was responsible for
spawning toast messages while copying a to-do to today.

This gave me the opportunity to improve and simplify a few things about
that system. The new logic should be easier to understand and maintain.
I'd not rule out that there are still some bugs, though.

Additionally, this commit adds a new `left_mouse_button` input action
which makes mouse button input checks a lot less verbose.

Also, adds a `release_focus` method to `Utils`, which can be used to
easily release whatever UI element is currently focused.