Fix calendar widget not properly closing
... when toggling the calendar widget button via a mouse click while the
calendar widget was already open. It would briefly hide and then show up
again, since the Overlay's `_input` handler operates on `action_pressed`
and hides it and the button operates on `action_released` and shows it.

Now, the Overlay's `_input` handler will only trigger after making sure
the mouse cursor is not over the calendar button. I'm not a huge fan of
this solution, but it's the best I could come up with...

Using `_unhandled_input` instead is *not* possible, as we e.g. also want
the widget to close when clicking a to-do list or unrelated button.
Using `action_released` in the Overlay is also not possible, as we also
want to hide the widget while the user starts dragging a to-do. But even
if we *did* change this, that would only stop the "flickering" of the
widget: Clicking the button would still not close the widget.