Fix calendar widget not matching the light theme
Since the Overlay isn't a Control node (it's a CanvasLayer), the theme
of the main window isn't inherited automatically, and the CalendarWidget
always used its dark variant, regardless of the current theme.

To fix that issue, we now manually sync the widgets theme with the main
window theme whenever the `dark_mode` setting is changed.