Undo `Utils.release_focus` method added in 8181915
... since I've now realized there's already a built-in method for that:
`get_viewport().set_input_as_handled()`.

However, we can't use the same function in `main_window.gd` (which is
what I was looking into originally), as there we want to return early
if there's already a focused Control in the viewport.

Also, fixes a bug in `todo_item.gd` that resulted in releasing the focus
of the currently edited to-do and resetting to the `previous_view_mode`
at the same time, when pressing `ui_cancel` (which is not wanted!).