Rework `Utils.is_mouse_cursor_above`
... to make use of `get_viewport().gui_get_hovered_control`. This makes
the method more precise: It's not just checking if the cursor is in the
correct region of the window, but if it's *actually* hovering it.
However, it also means the method can now only applied to Control nodes,
not just any node in the tree, which doesn't really matter in our case,
though, as the method is only used with Control nodes so far.

Also, limits the Overlay's `_input` method to only run through while the
CalendarWidget is visible (i.e. it won't consider clicks then).