Ignore to-do inputs while modifier key is held
When e.g. pressing CTRL+H (which is not mapped to anything in this app,
and also not used by Godot's LineEdit node) while editing a to-do, this
will result in the letter "h" getting added to the to-dos text.

This commit fixes that by automatically accepting any `gui_input` event
where a modifier key (either CTRL, ALT or META) is pressed, *unless* it
overlaps with the built-in keybinding of Godot's LineEdit node: CTRL+A,
CTRL+C, CTRL+X, CTRL+V, and CTRL+Z.
Ideally, I'd want this to be the default for *any* LineEdit in the app.
However, that is more involved, as some of Godot's built-in nodes (like
the SpinBox) internally use a LineEdit node as well.