Add method to check if event matches *any* action
That is, it loops over the entire InputMap and checks if the given event
matches any of the actions defined. As soon as it matches something, the
method returns true. If it doesn't match anything, it returns false.

I wish there was a built-in method to do this... (May open up an issue)
Also, would be lovely if `InputMap.get_actions` could be made ignoring
built-in actions via an argument, instead of filtering its output.

This method was required to fix a bug, where *any* keyboard input while
the CalendarWidget was open would hide it again, as long as that input
was not equal to the `toggle_calendar_widget` action. Now, the method
will only hide the widget, if the event matches another input action.
That is, just hitting CTRL for example will *not* close it.