Set all Settings UI Elements to `FOCUS_NONE`
For the Buttons and CheckButtons this was already the case. After this
commit it's also true for the remaining UI elements! In the case of the
SpinBox, that also required a change in `better_spin_box.gd`: By default
pressing ESCAPE will only leave the internal line edit's edit mode, but
keeps it focused nonetheless, even if the focus mode is `FOCUS_NONE`.

I'm aware that being able to focus and navigate between UI elements with
the keyboard is an important accesibility feature. It's just that during
active development this is very prone to breaking and behaving different
from what you'd expect, thus I prefer to disable it completely for now.
Once the app will be in a more stable state, I'll do it properly!