Add some minor improvements to the BetterSpinBox
First, makes the SpinBox ignore both middle and right mouse button
presses. That means it's no longer possible to focus the SpinBox this
way, paste text via the middle mouse button, or open the line edit's
context menu via the right mouse button.

Secondly, forces the (internal) LineEdit node to release the focus on
`text_submitted` if the focus mode is set to `FOCUS_NONE` (analogous to
what already was the case when pressing `ui_cancel`).