Block any non-numerical inputs on BetterSpinBox
That is, it's only possible to enter numbers between 0 and 9 (either via
the regular numbers row or the numpad), and one period character (unless
the `format_string` ends with a "d", i.e. enforces an integer).

The node will refuse to add more than one period character.

Using the comma key (e.g. on German keyboards) remains possible, but the
character will be immediately replaced by a period (which equals Godot's
default behavior when submitting the text).

This means it's no longer possible to do any arithmetic operations (e.g.
dividing a number by another one or adding two values) with the SpinBox.
The previous behavior can be reinstantiated by setting a BetterSpinBox'
`numerical_input_only` export variable to false, if that's required.

Special care was taken to *not* block any built-in user actions, though.
To that effect, a new `Utils.is_built_in_action` method was added.