Limit the range of valid BetterSpinBox inputs
When attempting to change the (internal) line edit's text to a value
that's below the SpinBox' `min_value` or above its `max_value`, the
change will get autocorrected to the respective limit.

Also, the line edit's `max_length` is now capped at the length of the
`max_value` (after being formatted according to `format_string`).