Add a better way to ignore modifier inputs
The old approach used a whitelist (which happened to be incomplete as
well: it broke the "ui_text_caret_word_left" action, for example).
The new approach hooks into `_unhandled_key_input` and is much simpler:
It only triggers when pressing a keycode that corresponds to an Unicode
character and after all input actions have been marked as handled.

Changes apply to `BetterLineEdit`, `BetterSpinBox` and `BetterTextEdit`.
It can be disabled via an `ignore_modifier_inputs` export variable.