Fix some smaller bugs with custom tooltips
Previously, pressing any key while a tooltip was shown would hide the
tooltip and it would not show up again until the user moved the mouse
off the host and back on to it. Starting with this commit, that is no
longer the case: Once a key is released, the tooltip is re-triggered.
Note that no tooltip will show until *all* keys are released!

This commit also gets rid of the `_on_press_or_toggle` callback, which
is no longer required as everything is handled via the `_input` method
now which only stays enabled while the host is hovered with the mouse.
To reduce the load even further, echo-events will be ignored.

Also, applies the project's script template to `tooltip.gd` by adding
both a `_set_initial_state` and `_connect_signals` method.