Add a custom `BetterButton` class
... that enforces updating the mouse cursor after the button is pressed.
Normally this only happens once the user actively moves the mouse, which
lead to issues with the user side panels as well as the day buttons in
the calendar widget, as both change what UI elements will be underneath
the cursor after pressing the button. For the side panel buttons it even
meant their tooltips (wrongly!) reappeared after a press.