Load to-do list data from the Data singleton
... instead of the `Cache` singleton, which eventually shall be phased
out. This commit leaves the previous system mostly intact for now, thus
hopefully shouldn't break anything. The main difference is that data is
no longer loaded from a string, but a full *object* representing it.

Note that this is (and for the foreseeable future likely will be) a work
in progress. Everything else is still done through the old system!

Also, adds a `_data` suffix to all classes used in `Data` for clarity.