Correctly restore to-do list's scroll offset
This is basically just migrating a part of the old `load_from_cache`
method to a property of `file_data.gd`. When `load_data` is called,
first we restore all items, then wait one process frame to give the
container time to update its size and then apply the scroll offset.

Also, fixes a line in `bookmark.gd` that still relied on the old
`States` enum in `todo_item.gd` which I missed in 33175c2.