Replace `_start_debounce_timer` calls
... by setting the respective variable in the corresponding ToDoData,
which will result in the `changed` signal being emitted and propagated
up until it reaches the `Data` singleton, that will create a new timer
or restart the existing one, if there is any.

While doing this, I fixed a number of bugs and oversights:
- Timers were always newly created, never restarted: As the timers are
  named after the corresponding filename (including the extension!) and
  node names in Godot aren't allowed to contain a dot, it got replaced
  by an underscore, which was not considered when looking up existing
  timers. To prevent this, the extension is now removed from the name!
- Pending timers weren't fast-forwarded (i.e. nothing was saved) when
  closing Habituary, only when the app got unfocused by the user
- The `changed` signal of any `sub_items` weren't forwarded to the layer
  above, making it impossible to save moficiations to sub items.

A few new DEBUG log messages were added for clarity, indicating when a
debounce timer is started (or restarted) and for what file, as well as
when a file is saved to disk (since its debounce timer timed out).

Also, fixes yet *another* line that I missed in 33175c2, which still
relied on the old `States` enum in `todo_item.gd`...