Log reason for (re)starting a debounce timer
By requiring a String argument providing the reason when emitting the
`changed` signal in either ToDoData or ToDoListData.

Also, adds a `shorten_string` method to `Utils`, which can be used to
limit potentially long strings to a certain width before printing them.
If the string is longer, its tail is replaced by the given placeholder
("..." by default), otherwise it's returned as is.