Remove DebounceTimer in `todo_list.gd` completely
That is, writing changes to any to-do to disk should now exclusively
happen through the `Data` singleton and nowhere else.

This includes adding an `unindent` method in `to_do_list_data.gd` which
required adding a `parent` variable to all `ToDoData` and `ToDoListData`
objects that points to the object hierarchically above them.